Backend/requirements.txt
aniso8601==7.0.0
APScheduler==3.10.4
bidict==0.23.1
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Faker==27.0.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-GraphQL==2.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Script==2.0.6
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
graphene==2.1.9
graphene-sqlalchemy==2.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
gunicorn==23.0.0
h11==0.14.0
importlib_metadata==8.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
mysql-connector-python==9.0.0
ordered-set==4.1.0
packaging==24.1
promise==2.3
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-http-client==3.3.7
python-socketio==5.11.3
pytz==2024.2
PyYAML==6.0.2
rich==13.7.1
Rx==1.6.3
sendgrid==6.11.0
simple-websocket==1.0.0
singledispatch==4.1.0
six==1.16.0
SQLAlchemy==2.0.32
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.0
